fix as per review
