+++ src/libs/ReportUtils.ts
@@ -8433,14 +8433,6 @@ function hasInvoiceReports() {
     return allReports.some((report) => isInvoiceReport(report));
 }
 
-function areAllGroupPoliciesExpenseChatDisabled(policies = allPolicies) {
-    const groupPolicies = Object.values(policies ?? {}).filter((policy) => isGroupPolicy(policy?.type ?? ''));
-    if (groupPolicies.length === 0) {
-        return false;
-    }
-    return !groupPolicies.some((policy) => !!policy?.isPolicyExpenseChatEnabled);
-}
-
 export {
     addDomainToShortMention,
     completeShortMention,
@@ -8768,7 +8760,6 @@ export {
     getAllReportErrors,
     getAllReportActionsErrorsAndReportActionThatRequiresAttention,
     hasInvoiceReports,
-    areAllGroupPoliciesExpenseChatDisabled,
 };
 
 export type {
